Reintroduce backendutil.TransformBackend after 1a1363a5da
Considering the comments https://github.com/emersion/go-smtp/issues/22
I still consider this backend to be useful.

The required state to keep track of from and to addresses can be
handled by the implementer by using struct bound Transform methods.
The TransformReset func will allow that state to be cleared.