server: Implement per-recipient status for LMTP
A separate interface is used so backend implementation will be aware
of LMTP usage. This also avoids clutter in simple backend
implementations that don't support LMTP.

RFC 2033 (Section 4.2) says that we should send multiple status lines if
the same address was specified multiple times. We move responsibility of
doing this to the backend implementation and don't do any deduplication
themselves. This allows transparent forwarding for LMTP to be
implemented correctly.